Rename collection table component
