+++ client/src/components/CollectionTable.vue
